service test first steps

+++ src/main/java/br/com/phoebus/microservice/biblioteca/loan/libraryloan/service/DeleteLibraryLoanServiceImpl.java
@@ -20,11 +20,14 @@ public class DeleteLibraryLoanServiceImpl implements DeleteLibraryLoanService {
 
     @Override//Ver se existe livros ainda em emprestimo nesse emprestimo, caso tenha, não excluir.
     public void deleteLibraryLoan(Long id) {
+        if(!libraryLoanRepository.existsById(id)) {
+            throw new LibraryLoanNotFoundException();
+        }
         List<LibraryBookDTO> libraryBookDTOList = userAndBookService.findAllBookOfLoan(id);
         if (libraryBookDTOList.size() > 0) {
             throw new LoanWithPendingBooksException();
         }
-        LibraryLoan libraryLoan = libraryLoanRepository.findById(id).orElseThrow(LibraryLoanNotFoundException::new);
+        LibraryLoan libraryLoan = libraryLoanRepository.getOne(id);
         userAndBookService.editUserSpecifId(libraryLoan.getSpecificIDUser(), "null");
         libraryLoanRepository.deleteById(id);
     }
